Use separate directories for building and running
Uses a separate directory for building the docker image and running
the other workflows. This is because the user the is used by the docker
git image is root which results in permission issues when the workflows
try to run. This change means that the directories do not overlap,
avoiding any permission issues.

Co-Authored-By: Will Szumski <will@stackhpc.com>