fix!: backwards compatibility with nested vars
Identified in `ansible-core < 2.15` nested vars will not be evaluated
within templates that are loaded via `extends` or `includes`. The
solution involves using `lookup('template', ...)` to ensure support
dating back to `2.12`. Unfortunately `lookup` does not have access to
the macro defined in `workflows.yml` nor does it inherit the
`block_*_string` and `variable_*_string` from the
`ansible.builtin.template` module as defined within `tasks/main.yml`.
Also `lookup` does not support setting `block_*_string``

NOTE: this is not ideal due to the duplication and the changing of
block/variable markers. Whilst low priority should be improved/replaced
in the future.